requirements.txt
python-pptx==0.6.21
Pillow==10.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2
